# expressjs-server/docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "ExpressJS REST API",
    "description": "<h4>Documentation automatically generated by <b>swagger-autogen</b></h4>"
  },
  "host": "localhost:5000",
  "basePath": "/",
  "tags": [
    {
      "name": "Home",
      "description": "Home page of application"
    },
    {
      "name": "Users",
      "description": "Users API"
    },
    {
      "name": "Permissions",
      "description": "Permissions API"
    },
    {
      "name": "Roles",
      "description": "Roles API"
    },
    {
      "name": "Login",
      "description": "Login Page"
    }
  ],
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/": {
      "get": {
        "tags": [
          "Home"
        ],
        "description": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/login": {
      "post": {
        "tags": [
          "Login"
        ],
        "description": "",
        "parameters": [
          {
            "name": "Login body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Login"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/users/active=active&keyWord=keyWord": {
      "get": {
        "tags": [
          "Users"
        ],
        "description": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "active",
            "in": "path",
            "description": "Filter by property: active(true, false, null)",
            "allowEmptyValue": true,
            "default": "",
            "type": "boolean",
            "required": true
          },
          {
            "name": "keyWord",
            "in": "path",
            "description": "Search by property: username",
            "allowEmptyValue": true,
            "default": "",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/users/": {
      "post": {
        "tags": [
          "Users"
        ],
        "description": "",
        "parameters": [
          {
            "name": "User body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Resource created succesfully"
          }
        }
      }
    },
    "/api/users/{id}": {
      "put": {
        "tags": [
          "Users"
        ],
        "description": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the user",
            "allowEmptyValue": false,
            "default": 0,
            "type": "integer",
            "required": true
          },
          {
            "name": "User body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Resource updated successfully"
          }
        }
      },
      "delete": {
        "tags": [
          "Users"
        ],
        "description": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the users",
            "allowEmptyValue": false,
            "default": 0,
            "type": "integer",
            "required": true
          }
        ],
        "responses": {
          "204": {
            "description": "Resource deleted successfully"
          }
        }
      }
    },
    "/api/roles/active=active&keyWord=keyWord": {
      "get": {
        "tags": [
          "Roles"
        ],
        "description": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "active",
            "in": "path",
            "description": "Filter by property: active(true, false, null)",
            "allowEmptyValue": true,
            "default": "",
            "type": "boolean",
            "required": true
          },
          {
            "name": "keyWord",
            "in": "path",
            "description": "Search by property: name",
            "allowEmptyValue": true,
            "default": "",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/roles/": {
      "post": {
        "tags": [
          "Roles"
        ],
        "description": "",
        "parameters": [
          {
            "name": "Role body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Role"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Resource created succesfully"
          }
        }
      }
    },
    "/api/roles/{id}": {
      "put": {
        "tags": [
          "Roles"
        ],
        "description": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the role",
            "allowEmptyValue": false,
            "default": 0,
            "type": "integer",
            "required": true
          },
          {
            "name": "Role body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Role"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Resource updated successfully"
          }
        }
      },
      "delete": {
        "tags": [
          "Roles"
        ],
        "description": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "204": {
            "description": "Resource deleted successfully"
          }
        }
      }
    },
    "/api/permissions/get": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "roleList": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  },
  "definitions": {
    "User": {
      "type": "object",
      "properties": {
        "username": {
          "type": "string",
          "example": "default_username"
        },
        "password": {
          "type": "string",
          "example": "password"
        },
        "active": {
          "type": "boolean",
          "example": true
        }
      }
    },
    "Permission": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "default_name"
        },
        "description": {
          "type": "string",
          "example": "Permission test description..."
        },
        "active": {
          "type": "boolean",
          "example": true
        }
      }
    },
    "Role": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "default_name"
        },
        "description": {
          "type": "string",
          "example": "Role test description..."
        },
        "active": {
          "type": "boolean",
          "example": true
        }
      }
    },
    "Login": {
      "type": "object",
      "properties": {
        "username": {
          "type": "string",
          "example": "user1"
        },
        "password": {
          "type": "string",
          "example": "pass"
        }
      }
    }
  }
}